+++ src/batch_selection_ui.ts
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import '@material/mwc-fab';
 import '@material/mwc-button';
 import './filtered_images_ui';
 import './filters_ui';
@@ -31,9 +30,7 @@ import {State} from './state';
 @customElement('batch-selection-ui')
 export class BatchSelectionUi extends LitElement {
   @property({attribute: false}) state!: State;
-
-  /** The selected batch. */
-  private batchSelection: BatchSelection|undefined = undefined;
+  @property({attribute: false}) batchSelection!: BatchSelection;
 
   @query('filtered-images-ui')
   private readonly filteredImagesUi: FilteredImagesUi|undefined;
@@ -43,123 +40,34 @@ export class BatchSelectionUi extends LitElement {
     listen(EventType.MATCHED_DATA_POINTS_CHANGED, () => {
       this.filteredImagesUi?.requestUpdate();
     });
-    listen(EventType.FILTERED_DATA_INFO_REQUEST, (event) => {
-      this.batchSelection = this.state.batchSelections[event.detail.batchIndex];
-      this.style.display = 'block';
-      this.requestUpdate();
-    });
   }
 
   override render() {
-    if (!this.batchSelection) return html``;
-
-    const onClose = () => {
-      this.batchSelection = undefined;
-      this.style.display = 'none';
-      this.requestUpdate();
-    };
-
     const batchIndex = this.batchSelection.batch.index;
     const onBatchInfoRequest = () => {
       dispatch(EventType.BATCH_INFO_REQUEST, {batchIndex});
-      onClose();
     };
     const onMatchesInfoRequest = () => {
       dispatch(EventType.MATCHES_INFO_REQUEST, {batchIndex});
-      onClose();
     };
 
     return html`
-      <div id="background" @click=${onClose}></div>
-      <div id="dialog" @click=${(e: Event) => {
-      e.stopImmediatePropagation();
-    }}>
-        <div class="horizontalFlex">
-          <filters-ui .state=${this.state}
-            .batchSelection=${this.batchSelection}>
-          </filters-ui>
-          <filtered-images-ui .state=${this.state}
-            .batchSelection=${this.batchSelection}>
-          </filtered-images-ui>
-        </div>
-
-        <div class="buttons">
-          <mwc-button
-            raised
-            icon="info"
-            label="Show metadata"
-            title="Display the details of the batch"
-            @click=${onBatchInfoRequest}>
-          </mwc-button>
-
-          ${
-        batchIndex === this.state.referenceBatchSelectionIndex ?
-            // disabled mwc-button title does not appear. Use a div.
-            html`
-          <div title="The reference batch cannot be matched with itself">
-            <mwc-button
-              raised
-              icon="join_inner"
-              label="Show matches"
-              disabled>
-            </mwc-button>
-          </div>
-        ` :
-            html`
-          <mwc-button
-            raised
-            icon="join_inner"
-            label="Show matches"
-            title="Display the matches between this batch and the reference batch"
-            @click=${onMatchesInfoRequest}>
-          </mwc-button>
-        `}
-        </div>
-
-        <mwc-fab id="closeButton" icon="close" title="Close" @click=${onClose}>
-        </mwc-fab>
+      <div class="horizontalFlex">
+        <filters-ui .state=${this.state} .batchSelection=${this.batchSelection}>
+        </filters-ui>
+        <filtered-images-ui .state=${this.state}
+          .batchSelection=${this.batchSelection}>
+        </filtered-images-ui>
       </div>`;
   }
 
   static override styles = css`
     :host {
-      display: none;
-      position: absolute;
-      z-index: 5;
-      left: 0;
-      top: 0;
-      bottom: 0;
-      right: 0;
-    }
-
-    #background {
-      width: 100%;
-      height: 100%;
-      background-color: rgba(0, 0, 0, 0.6);
-    }
-
-    #dialog {
-      background-color: var(--mdc-theme-surface);
-      position: absolute;
-      left: 40px;
-      top: 40px;
-      bottom: 40px;
-      right: 40px;
-      padding: 20px;
-      box-shadow: 0 0 8px 0 rgba(0, 0, 0, 0.2);
-      border-radius: 16px;
       display: flex;
       flex-direction: column;
-      /* Rely on margin:auto for distributing the space. */
       justify-content: flex-start;
       gap: 20px;
       overflow: hidden;
-    }
-
-    #closeButton {
-      position: absolute;
-      top: 20px;
-      right: 20px;
     }
 
     .horizontalFlex {
@@ -174,12 +82,6 @@ export class BatchSelectionUi extends LitElement {
       min-height: 150px;
       max-height: 100%;
       flex: 1;
-    }
-
-    .buttons {
-      display: flex;
-      flex-direction: row;
-      justify-content: space-evenly;
     }
   `;
 }